move effective_gas_price to utils
