+++ cmd/ef_tests/levm/runner/revm_runner.rs
@@ -4,11 +4,11 @@ use crate::{
         levm_runner::{self, post_state_root},
         EFTestRunnerError, InternalError,
     },
-    types::{EFTest, EFTestTransaction},
-    utils::load_initial_state,
+    types::EFTest,
+    utils::{effective_gas_price, load_initial_state},
 };
 use bytes::Bytes;
-use ethrex_core::{types::TxKind, Address, H256, U256};
+use ethrex_core::{types::TxKind, Address, H256};
 use ethrex_levm::{
     errors::{TransactionReport, TxResult},
     Account, StorageSlot,
@@ -88,19 +88,6 @@ pub fn re_run_failed_ef_test_tx(
         re_run_report,
     )?;
     Ok(())
-}
-
-// If gas price is not provided, calculate it with current base fee and priority fee
-pub fn effective_gas_price(test: &EFTest, tx: &&EFTestTransaction) -> U256 {
-    match tx.gas_price {
-        None => {
-            let current_base_fee = test.env.current_base_fee.unwrap();
-            let priority_fee = tx.max_priority_fee_per_gas.unwrap();
-            let max_fee_per_gas = tx.max_fee_per_gas.unwrap();
-            std::cmp::min(max_fee_per_gas, current_base_fee + priority_fee)
-        }
-        Some(price) => price,
-    }
 }
 
 pub fn prepare_revm_for_tx<'state>(
